Criacao do jacksonConfig.java

+++ src/main/java/br/com/teste/vander/config/Initialization.java
@@ -1,6 +1,5 @@
 package br.com.teste.vander.config;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Arrays;
 
@@ -60,9 +59,9 @@ public class Initialization implements CommandLineRunner {
 		Categoria c1 = new Categoria(null, "Informática");
 		Categoria c2 = new Categoria(null, "Escritório");
 		
-		Produto p1 = new Produto(null, "Computador", BigDecimal.valueOf(2000.00));
-		Produto p2 = new Produto(null, "Impressora", BigDecimal.valueOf(800.00));
-		Produto p3 = new Produto(null, "Mouse", BigDecimal.valueOf(80.00));
+		Produto p1 = new Produto(null, "Computador", 2000.00);
+		Produto p2 = new Produto(null, "Impressora", 800.00);
+		Produto p3 = new Produto(null, "Mouse", 80.00);
 
 		c1.getProdutos().addAll(Arrays.asList(p1,p2,p3));
 		c2.getProdutos().addAll(Arrays.asList(p2));
